Replace docsearch with instantsearch and algolia
