+++ assets/styles/_main_docsearch.scss
@@ -1,12 +1,11 @@
-@import "base/variables";
-@import "dropdown";
-@import "searchbox";
+@import 'base/variables';
+@import 'dropdown';
+@import 'searchbox';
 
-$searchbox: true!default;
+$searchbox: true !default;
 
-@if($searchbox == true){
-  @include searchbox($searchbox-config...);
+@if ($searchbox == true) {
+    @include searchbox($searchbox-config...);
 }
 
 @include dropdown($dropdown-config...);
-
