+++ assets/styles/style.scss
@@ -19,7 +19,7 @@
 @import 'vendor/chroma-styles';
 
 // docsearch
-@import 'main_docsearch';
+// @import 'main_docsearch';
 
 // global to all pages
 @import 'pages/global';
@@ -31,7 +31,7 @@
 // partial css lives with partial
 
 @import "pages/api";
-@import "pages/search";
+// @import "pages/search";
 @import "pages/404";
 @import "pages/home";
 @import "pages/glossary";
