+++ layouts/partials/home-header.html
@@ -2,41 +2,35 @@
 
 <!-- get lang specific data file -->
 {{ $.Scratch.Set "data" "" }}
-{{ if ne $dot.Page.Lang "en"}}
+{{ if ne $dot.Page.Lang "en" }}
     {{ if (fileExists (print "data/partials/home." $dot.Page.Lang ".yaml")) }}
-      {{ $.Scratch.Set "data" (index $dot.Page.Site.Data.partials (print "home." $dot.Page.Lang)) }}
+        {{ $.Scratch.Set "data" (index $dot.Page.Site.Data.partials (print "home." $dot.Page.Lang)) }}
     {{ else }}
-      {{ $.Scratch.Set "data" $dot.Page.Site.Data.partials.home }}
+        {{ $.Scratch.Set "data" $dot.Page.Site.Data.partials.home }}
     {{ end }}
 {{ else }}
     {{ $.Scratch.Set "data" $dot.Page.Site.Data.partials.home }}
 {{ end }}
 {{ $datafile := ($.Scratch.Get "data") }}
 
-
 <section class="home-hero">
     <div class="container">
         <div class="row">
             <div class="col-12 offset-0 col-md-8 offset-md-2">
                 <h1 class="mt-3 mt-md-8 mb-0 mb-md-1 text-white font-regular text-center">{{ $datafile.heading }}</h1>
-                <div class="search-container">
-                        <form class="docssearch mt-0" method="GET" action="{{ "search/" | absLangURL }}">
-                            <div class="input-group" id="docssearchWrapper">
-                                <input class="docssearch-input form-control" name="s" type="docssearch" placeholder="Search documentation...">
-                                <div class="input-group-append">
-                                    <button class="btn btn-lg btn-primary text-uppercase js-search-btn" type="button" role="button">
-                                        <span class="d-none d-md-flex">Search</span>
-                                        <span class="d-flex d-md-none">{{ partial "svg" (dict "context" $dot "src" "search.svg" "color" "white" )}}</span>
-                                    </button>
-                                </div>
-                            </div>
-                        </form>
+
+                <div class="searchbox-container">
+                    <div id="searchbox"></div>
+                    <div>
+                        <div id="hits" class="d-none"></div>
+                    </div>
                 </div>
+
                 <p class="smaller text-white text-center font-semibold popular-search">
                     Popular searches include
-                    {{ $lenList := (len $datafile.popular_searches)}}
+                    {{ $lenList := (len $datafile.popular_searches) }}
                     {{ range $index, $popular_search := (sort $datafile.popular_searches "weight" "asc") }}
-                        <a href="{{ $popular_search.link | absLangURL }}">{{ $popular_search.title }}</a>{{ if eq (sub $lenList 2) $index }}, and{{ else if eq (sub $lenList 1) $index }}{{ else }},{{ end}}
+                        <a href="{{ $popular_search.link | absLangURL }}">{{ $popular_search.title }}</a>{{ if eq (sub $lenList 2) $index }}, and {{ else if eq (sub $lenList 1) $index }}{{ else }},{{ end }}
                     {{ end }}
                 </p>
             </div>
