+++ layouts/partials/search-mobile.html
@@ -1,4 +1,4 @@
-<form class="docssearch" method="GET" action="{{ "search/" | absLangURL }}" style="width:90%">
+{{/*  <form class="docssearch" method="GET" action="{{ "search/" | absLangURL }}" style="width:90%">
     <div class="search-input-container">
         <svg class="search-input-icon" width="18" height="19" viewBox="0 0 13 14" fill="none"
           xmlns="http://www.w3.org/2000/svg">
@@ -10,4 +10,4 @@
           autocomplete="off" />
         <label class="sr-only" for="search-input-mobile">Search Docs</label>
       </div>
-</form>
+</form>  */}}
