+++ layouts/partials/search.html
@@ -1,4 +1,4 @@
-<form class="docssearch" method="GET" action="{{ "search/" | absLangURL }}">
+{{/*  <form class="docssearch" method="GET" action="{{ "search/" | absLangURL }}">
   <div class="search-input-container">
     <svg class="search-input-icon" width="13" height="14" viewBox="0 0 13 14" fill="none"
       xmlns="http://www.w3.org/2000/svg">
@@ -12,4 +12,4 @@
     <label class="sr-only" for="search-input">Search Docs</label>
 
   </div>
-</form>
+</form>  */}}
