+++ layouts/partials/sidenav/api-sidenav.html
@@ -5,7 +5,7 @@
             <a href="{{ "" | absLangURL  }}">
                 <p class="h4">{{ i18n "datadog_docs" }} &nbsp;&nbsp;<span class="font-semibold">API</span></p>
             </a>
-            <form class="docssearch" method="GET" action="{{ "search/" | absLangURL }}">
+            {{/*  <form class="docssearch" method="GET" action="{{ "search/" | absLangURL }}">
                 <div class="search-input-container">
                     <svg class="search-input-icon" width="13" height="14" viewBox="0 0 13 14" fill="none" xmlns="http://www.w3.org/2000/svg">
                         <path fill-rule="evenodd" clip-rule="evenodd" d="M8.88014 6.04741C8.88014 5.13185 8.5557 4.34963 7.9068 3.70074C7.25791 3.05185 6.47125 2.72296 5.5468 2.71407C4.62235 2.70518 3.8357 3.03407 3.1868 3.70074C2.53791 4.36741 2.21347 5.14962 2.21347 6.04741C2.21347 6.94519 2.53791 7.73185 3.1868 8.4074C3.8357 9.08296 4.62235 9.4074 5.5468 9.38074C6.47125 9.35407 7.25791 9.02963 7.9068 8.4074C8.5557 7.78518 8.88014 6.99852 8.88014 6.04741ZM12.6935 12.2474C12.6935 12.5052 12.5957 12.7274 12.4001 12.9141C12.2046 13.1008 11.9779 13.1941 11.7201 13.1941C11.4624 13.1941 11.2446 13.1008 11.0668 12.9141L8.52013 10.3674C7.63124 10.9808 6.64014 11.2874 5.5468 11.2874C4.83569 11.2874 4.15569 11.1497 3.5068 10.8741C2.85791 10.5985 2.30236 10.2252 1.84013 9.7541C1.37791 9.28299 1.00458 8.72744 0.720134 8.08743C0.435688 7.44743 0.297912 6.76744 0.306801 6.04743C0.315689 5.32743 0.453466 4.65188 0.720134 4.02077C0.986802 3.38965 1.36013 2.82966 1.84013 2.34077C2.32014 1.85187 2.87569 1.47854 3.5068 1.22077C4.13791 0.962986 4.81791 0.82521 5.5468 0.807432C6.27569 0.789654 6.95569 0.927431 7.5868 1.22077C8.21792 1.5141 8.77347 1.88743 9.25347 2.34077C9.73347 2.7941 10.1068 3.3541 10.3735 4.02077C10.6401 4.68744 10.7779 5.36298 10.7868 6.04743C10.7868 7.14077 10.4801 8.13187 9.8668 9.02077L12.4135 11.5674C12.6001 11.7541 12.6935 11.9808 12.6935 12.2474Z" fill="#B4B1B6"/>
@@ -16,7 +16,7 @@
                     <label class="sr-only" for="search-input-api">Search Docs</label>
 
                 </div>
-            </form>
+            </form>  */}}
             <div class="sidenav-nav">
 
                 {{ partial "nav/left-nav-api.html" .}}
