+++ package.json
@@ -44,12 +44,12 @@
         "@datadog/browser-logs": "^4.19.1",
         "@datadog/browser-rum": "^4.19.1",
         "a11y": "https://s3.amazonaws.com/origin-static-assets/corp-node-packages/master/a11y-v1.0.1.tgz",
-        "algoliasearch": "4.11.0",
+        "algoliasearch": "^4.14.2",
         "bootstrap": "4.3.1",
         "del": "4.1.1",
-        "docsearch.js": "^2.6.3",
         "fancy-log": "^1.3.3",
         "hugo-bin": "0.93.0",
+        "instantsearch.js": "^4.49.1",
         "jquery": "3.5.1",
         "js-cookie": "^2.2.1",
         "js-yaml": "^3.14.1",
